--- feature/src/main/java/com/kotlin/feature/search/SearchFragment.kt
@@ -1,12 +1,0 @@
-package com.kotlin.feature.search
-
-import com.kotlin.feature.R
-import com.kotlin.feature.databinding.FragmentSearchBinding
-import com.sopt.now.core.base.BindingFragment
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class SearchFragment : BindingFragment<FragmentSearchBinding>(R.layout.fragment_search) {
-    override fun initView() {
-    }
-}
