+++ feature/src/main/java/com/sopt/now/feature/home/HomeDetailFragment.kt
@@ -1,0 +1,39 @@
+package com.sopt.now.feature.home
+
+import android.os.Build
+import androidx.core.view.isVisible
+import com.now.feature.R
+import com.now.feature.databinding.FragmentMyPageBinding
+import com.sopt.now.feature.model.HomeSealedItem
+import com.sopt.now.feature.util.KeyStorage
+import com.sopt.now.core.base.BindingFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class HomeDetailFragment : BindingFragment<FragmentMyPageBinding>(R.layout.fragment_my_page) {
+    override fun initView() {
+        updateHomeDetailUI(initGetFriendData())
+    }
+
+    private fun initGetFriendData(): HomeSealedItem.Friend? {
+        return arguments?.let {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                it.getParcelable(KeyStorage.USER_INPUT, HomeSealedItem.Friend::class.java)
+            } else {
+                it.getParcelable(KeyStorage.USER_INPUT)
+            }
+        }
+    }
+
+    private fun updateHomeDetailUI(friend: HomeSealedItem.Friend?) = with(binding) {
+        friend?.let {
+            ivMainProfile.setImageResource(friend.profileImage)
+            tvMainIdData.text = friend.name
+            tvMainNicknameData.text = friend.name
+            tvMainMbtiData.text = friend.date.toString()
+            tvMainPwdData.text = friend.selfDescription
+            tvMainSignOut.isVisible = false
+            tvMainClearInfo.isVisible = false
+        }
+    }
+}
