+++ feature/src/main/java/com/sopt/now/feature/search/SearchFragment.kt
@@ -1,0 +1,12 @@
+package com.sopt.now.feature.search
+
+import com.now.feature.R
+import com.now.feature.databinding.FragmentSearchBinding
+import com.sopt.now.core.base.BindingFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class SearchFragment : BindingFragment<FragmentSearchBinding>(R.layout.fragment_search) {
+    override fun initView() {
+    }
+}
